Throw error when a promise is rejected
This change makes it possible for client code to distinguish between rejected
and failed promises when attaching then, or catch handlers.  With the previous
behaviour it was impossible to make this distinction (at least using the
promise API).

If the old behaviour is desired, a client can add a catch handler which
returns the error before dispatching:

```
dispatch(myPromise.catch((err) => err))
```