Don't clobber action meta
Previously, if an action had both meta data and payload, the metadata would be
clobbered with the payload once the promise had either been rejected or
resolved.

Now, the original payload is added to the metadata under the key `payload`
preserving any meta already present; unless there was meta under the key
`payload` which seems unlikely.

This is a breaking API change.